+++ CHANGELOG.md
@@ -5,6 +5,9 @@ log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- Showing an update modal before updating a record.
+
 ## [0.30.1] - 2025-09-07
 ### Added
 - Support for SQL mode.